requirements.txt
pydantic>=2.0
openai>=1.40
anthropic>=0.30
httpx>=0.27
pyyaml>=6.0
python-dotenv>=1.0
GitPython>=3.1
rich>=13.7